+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Added definition of 8k resolution
  - #34: Now rejects unsupported features in text inputs
  - #30: Fixed cloud upload for VOD
+ - #29: Added webcam support on macOS
 
 
 ## 0.2.0 (2019-10-14)